backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
numpy
pandas
scikit-learn
xgboost
joblib
firebase-admin
razorpay
